Moved floorplan to presentational
