+++ client/src/components/Presentational/FloorPlan/styles.js
